Move CPULoadStoreSpec into common fie
To avoid fragile circular dependencies woith manager, moving
CPULoadStoreSpec into common file.

Review comment:

https://github.com/vllm-project/vllm/pull/40538#discussion_r3128587928
