+++ tests/v1/kv_offload/cpu/test_gpu_worker.py
@@ -15,9 +15,9 @@ from vllm.v1.kv_offload.base import (
     CanonicalKVCacheTensor,
     GPULoadStoreSpec,
 )
+from vllm.v1.kv_offload.cpu.common import CPULoadStoreSpec
 from vllm.v1.kv_offload.cpu.gpu_worker import CpuGpuOffloadingHandlers
 from vllm.v1.kv_offload.cpu.shared_offload_region import SharedOffloadRegion
-from vllm.v1.kv_offload.cpu.spec import CPULoadStoreSpec
 
 NUM_GPU_BLOCKS = [64]
 NUM_CPU_BLOCKS = [256]
