+++ vllm/v1/kv_offload/cpu/spec.py
@@ -6,25 +6,19 @@ from vllm.config import VllmConfig
 from vllm.platforms import current_platform
 from vllm.v1.kv_cache_interface import KVCacheConfig
 from vllm.v1.kv_offload.base import (
-    BlockIDsLoadStoreSpec,
     CanonicalKVCaches,
     GPULoadStoreSpec,
     LoadStoreSpec,
     OffloadingManager,
     OffloadingSpec,
 )
+from vllm.v1.kv_offload.cpu.common import CPULoadStoreSpec
 from vllm.v1.kv_offload.cpu.gpu_worker import CpuGpuOffloadingHandlers
+from vllm.v1.kv_offload.cpu.manager import (
+    CPUOffloadingManager,
+    FilterReusedOffloadingManager,
+)
 from vllm.v1.kv_offload.worker.worker import OffloadingHandler
-
-
-class CPULoadStoreSpec(BlockIDsLoadStoreSpec):
-    """
-    Spec for loading/storing a KV block to CPU memory.
-    """
-
-    @staticmethod
-    def medium() -> str:
-        return "CPU"
 
 
 class CPUOffloadingSpec(OffloadingSpec):
@@ -64,11 +58,6 @@ class CPUOffloadingSpec(OffloadingSpec):
 
     def get_manager(self) -> OffloadingManager:
         if not self._manager:
-            from vllm.v1.kv_offload.cpu.manager import (
-                CPUOffloadingManager,
-                FilterReusedOffloadingManager,
-            )
-
             kv_events_config = self.vllm_config.kv_events_config
             enable_events = (
                 kv_events_config is not None and kv_events_config.enable_kv_cache_events
